Move attribute into utilities
