+++ lib/utilities/attribute.js
@@ -1,10 +1,10 @@
 // Highly modeled after the interface for YUI's Attribute.
 // Though much simpler internally. 
 
-var ObjectUtils = require('./utilities/objects');
-var Lang = require('./utilities/lang');
-var EventTarget = require('./eventTarget');
-var augment = require('./utilities/augment');
+var ObjectUtils = require('./objects');
+var Lang = require('./lang');
+var EventTarget = require('./../eventTarget');
+var augment = require('./augment');
 
 var OPTS = {
   GETTER: 'getter',
